backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary  
pydantic-settings 
alembic          
python-jose[cryptography]
passlib[bcrypt]  
python-multipart 
argon2-cffi
fastapi-mail